Moveing to a central meta data class
